Fixed warnings. JIRA: CXX-4704
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@63784 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/data_loaders/genbank/dispatcher.cpp
@@ -1510,7 +1510,7 @@ void CReadDispatcher::LoadBlobSet(CReaderRequestResult& result,
 }
 
 
-bool CReadDispatcher::SetBlobState(int i,
+bool CReadDispatcher::SetBlobState(size_t i,
                                    CReaderRequestResult& result,
                                    const TIds& ids, TLoaded& loaded,
                                    TStates& ret)
